requirements.txt
PyYAML
Pandas
beautifulsoup4
requests
lxml